Querying ranges by score in case of non string values

+++ Redis.SQL.Client/Enums/RedisBoolean.cs
@@ -1,6 +1,6 @@
 ﻿namespace Redis.SQL.Client.Enums
 {
-    internal enum RedisBoolean : long
+    internal enum RedisBoolean : int
     {
         False = 0,
         True = 1
